added versioning and dto

+++ backend/DTO/Common/NotificationDTO.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 namespace backend.DTO.Common;
 
 public record NotificationDto(
-    int NotificationId,
+    long NotificationId,
     string Title,
     string Message,
     bool IsRead,
